Check more rigorously that API proxy launched successfully
This fixes a bug that causes vessel init to fail opaquely when it fails to launch the fly machine api-proxy command.

The previous implementation assumed that the api-proxy command succeeded as long as the command launches successfully, but that meant it wouldn't notice if the API proxy died with an error before it began serving.

This change adjusts the StartMachineProxy function so that it waits to return until it verifies a successful TCP connection to the API proxy. If it can't connect within 2 seconds, it attempts to return the stderr from the command as an error message.

This change also fixes related bugs in init.go and destroy.go where the callers were accidentally discarding the error return value from StartMachineProxy.