Fix error propagation mistake when testing SSH connection
This fixes a bug where we were accidentally swallowing the error from TestConnection instead of printing it to the user. We were defining a new err variable with limited scope, but then later on, we tried to print its value even though at the point of printing, err referred to a different variable that was always nil.

This refactors the init code so that we print the error to the terminal and makes it harder to confuse different err variables.