Fix segfault in http.go
There was a bug in the http package where if DoRequest exhausts all of its attempts, we still try to call result.Body.Close() even though result is nil.

This change refactors the fly/http.go file to eliminate the segfault and make it harder to accidentally use unintialized variables.